Fixing the Rubocop violations for the BitstreamExportJob
